Create arg_diff.boa
An attempt to match a method_call with its counterpart in the next version.

Still in progress.

Description of the bug: 

"Basically, if we have
Old version
foo(a, "b", c)
foo(a, "b", c)

And in the new one, we have
foo(m, "n", o)
foo(a, "b", c)

then we can't reliably detect which foo(a, "b", c) corresponds with (was changed into) foo(m, "n", o). 

It's possible that one of the foo(a, "b", c) was not changed at all. Intuitively, it is probably the second one. However, it is possible that another call to foo was added or removed before these two calls, which would cause a mismatch between the versions.

Our script can't tell (and I can't really think of a way off the top of my head) which foo(a, "b", c) was changed. Basically, a call to a method may completely match many calls in another version, and vice versa."